Standardized all #define guards and renamed ui/algorithm.*
