+++ alg/shapeformation.h
@@ -15,8 +15,8 @@
 // mode == "t2" --> center triangle formation
 // mode == "l" --> line formation
 
-#ifndef AMOEBOTSIM_ALG_SHAPEFORMATION_H
-#define AMOEBOTSIM_ALG_SHAPEFORMATION_H
+#ifndef AMOEBOTSIM_ALG_SHAPEFORMATION_H_
+#define AMOEBOTSIM_ALG_SHAPEFORMATION_H_
 
 #include "core/amoebotparticle.h"
 #include "core/amoebotsystem.h"
@@ -127,4 +127,4 @@ class ShapeFormationSystem : public AmoebotSystem  {
   static std::set<QString> getAcceptedModes();
 };
 
-#endif  // AMOEBOTSIM_ALG_SHAPEFORMATION_H
+#endif  // AMOEBOTSIM_ALG_SHAPEFORMATION_H_
