+++ core/amoebotsystem.h
@@ -5,8 +5,8 @@
 // Defines a particle system that complies with the properties and constraints
 // of the amoebot model, but is independent of any particular algorithm.
 
-#ifndef AMOEBOTSIM_ALG_AMOEBOTSYSTEM_H
-#define AMOEBOTSIM_ALG_AMOEBOTSYSTEM_H
+#ifndef AMOEBOTSIM_CORE_AMOEBOTSYSTEM_H_
+#define AMOEBOTSIM_CORE_AMOEBOTSYSTEM_H_
 
 #include <deque>
 #include <map>
@@ -87,4 +87,4 @@ class AmoebotSystem : public System, public RandomNumberGenerator {
   unsigned int _numRounds;
 };
 
-#endif  // AMOEBOTSIM_ALG_AMOEBOTSYSTEM_H
+#endif  // AMOEBOTSIM_CORE_AMOEBOTSYSTEM_H_
