+++ core/system.h
@@ -5,8 +5,8 @@
 // Defines a base particle system and its iterators, primarily serving the
 // purpose of defining functions to be overridden by subclasses.
 
-#ifndef AMOEBOTSIM_SIM_SYSTEM_H
-#define AMOEBOTSIM_SIM_SYSTEM_H
+#ifndef AMOEBOTSIM_CORE_SYSTEM_H_
+#define AMOEBOTSIM_CORE_SYSTEM_H_
 
 #include <QMutex>
 
@@ -114,4 +114,4 @@ bool System::isConnected(const ParticleContainer& particles) {
   return occupiedNodes.empty();
 }
 
-#endif  // AMOEBOTSIM_SIM_SYSTEM_H
+#endif  // AMOEBOTSIM_CORE_SYSTEM_H_
