+++ ui/algorithm.cpp
@@ -2,7 +2,7 @@
  * The full GNU GPLv3 can be found in the LICENSE file, and the full copyright
  * notice can be found at the top of main/main.cpp. */
 
-#include "ui/alg.h"
+#include "ui/algorithm.h"
 
 Algorithm::Algorithm(QString name, QString signature)
     : _name(name),
