+++ ui/commandhistorymanager.h
@@ -2,8 +2,8 @@
  * The full GNU GPLv3 can be found in the LICENSE file, and the full copyright
  * notice can be found at the top of main/main.cpp. */
 
-#ifndef COMMANDHISTORYMANAGER_H
-#define COMMANDHISTORYMANAGER_H
+#ifndef AMOEBOTSIM_UI_COMMANDHISTORYMANAGER_H_
+#define AMOEBOTSIM_UI_COMMANDHISTORYMANAGER_H_
 
 #include <deque>
 
@@ -29,4 +29,4 @@ protected:
     int currentIndex;
 };
 
-#endif // COMMANDHISTORYMANAGER_H
+#endif  // AMOEBOTSIM_UI_COMMANDHISTORYMANAGER_H_
